% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topPolluters.R
\docType{data}
\name{topPolluters}
\alias{topPolluters}
\title{Top Climate, Air, and Water Polluters (2018)}
\format{
A data frame with 220 observations on the following 8 variables:
\describe{
  \item{polluter}{Parent Entity or Corporation}
  \item{toxic.air.rank}{Toxic 100 Air Rank; the company's rank for air toxics 2018}
  \item{greenhouse.rank}{Greenhouse 100 Rank; the company's rank for greenhouse gases in 2018}
  \item{toxic.air.poor}{Toxic 100 Air EJ: Poor Share; shares of potential exposure to air toxics borne by people living below the poverty line}
  \item{toxic.air.minority}{Toxic 100 Air EJ: Minority Share; shares of potential exposure to air toxics borne by people in minority racial/ethnic groups}
  \item{toxic.water.rank}{Toxic 100 Water Rank; the company's rank for water toxics from large facilities in 2018}
  \item{greenhouse.poor}{Greenhouse 100 EJ: Poor Share; shares of potential exposure to co-pollutants of combustion borne by people living below the poverty line}
  \item{greenhouse.minority}{Greenhouse 100 EJ: Minority Share; shares of potential exposure to co-pollutants of combustion borne by people in minority racial/ethnic groups}
}
}
\source{
\url{https://www.peri.umass.edu/combined-toxic-100-greenhouse-100-indexes-current}
}
\usage{
topPolluters
}
\description{
Data on the combined toxic 100 / greenhouse 100 indexes (2020 report, based on 2018 data). Researchers at the Political Economy Research Institute (PERI) at the University of Massachusetts Amherst published new editions of the  Greenhouse 100  Index, ranking U.S. companies by their emissions responsible for global climate change according to the U.S. EPA Greenhouse Gas Reporting Program, and the Toxic 100 Air and Toxic 100 Water Indexes, ranking U.S. industrial polluters using the U.S. EPA Toxics Release Inventory. The PERI Indexes include Environmental Justice indicators to assess impacts on low-income people and minorities.
}
\keyword{datasets}
